requirements.txt
requests
beautifulsoup4
pywebview
tqdm
hachoir
httpx
aiohttp
pyinstaller
Pillow
